:recycle: Add controller layer to user routes
